feat: js to ts conversion
