+++ mobile/components/LoadingSpinner.tsx
@@ -1,0 +1,15 @@
+import { View, ActivityIndicator, StyleSheet } from "react-native";
+
+export const LoadingSpinner = () => (
+  <View style={styles.container}>
+    <ActivityIndicator size="large" color="#0000ff" />
+  </View>
+);
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+});
