+++ mobile/components/index.ts
