+++ mobile/firebase-config.ts
