+++ mobile/package.json
@@ -9,24 +9,25 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/webpack-config": "^19.0.0",
+    "@react-native-async-storage/async-storage": "1.18.2",
     "expo": "~49.0.15",
+    "expo-image": "~1.3.5",
     "expo-image-picker": "~14.3.2",
+    "expo-linear-gradient": "~12.3.0",
     "expo-status-bar": "~1.6.0",
+    "expo-updates": "~0.18.17",
     "firebase": "^10.6.0",
     "react": "18.2.0",
-    "react-native": "0.72.6",
-    "expo-image": "~1.3.5",
-    "react-native-web": "~0.19.6",
     "react-dom": "18.2.0",
-    "@expo/webpack-config": "^19.0.0",
-    "typescript": "^5.1.3",
-    "@types/react": "~18.2.14",
-    "expo-linear-gradient": "~12.3.0",
-    "expo-updates": "~0.18.17",
-    "@react-native-async-storage/async-storage": "1.18.2"
+    "react-native": "0.72.6",
+    "react-native-web": "~0.19.6",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
+    "@types/react": "^18.2.38",
+    "@types/react-native": "^0.72.7",
     "react-native-dotenv": "^3.4.9"
   },
   "private": true
