+++ mobile/types/env.d.ts
@@ -1,0 +1,8 @@
+declare module "@env" {
+  export const BASE_URL;
+  export const FIREBASE_API_KEY;
+  export const FIREBASE_STORAGE_BUCKET;
+  export const FIREBASE_APP_ID;
+  export const FIREBASE_PROJECT_ID;
+  export const FIREBASE_AUTH_DOMAIN;
+}
